create an interface for gh creds provider
We plan to build on this later to allow the credentials provider to
be passed into the scaffolder tasks, the processors, and the url
readers.
