+++ packages/backend-common/src/reading/GithubUrlReader.ts
@@ -16,6 +16,7 @@
 
 import {
   getGitHubFileFetchUrl,
+  DefaultGithubCredentialsProvider,
   GithubCredentialsProvider,
   GitHubIntegration,
   ScmIntegrations,
@@ -58,7 +59,7 @@ export class GithubUrlReader implements UrlReader {
   static factory: ReaderFactory = ({ config, treeResponseFactory }) => {
     const integrations = ScmIntegrations.fromConfig(config);
     return integrations.github.list().map(integration => {
-      const credentialsProvider = GithubCredentialsProvider.create(
+      const credentialsProvider = DefaultGithubCredentialsProvider.create(
         integration.config,
       );
       const reader = new GithubUrlReader(integration, {
